% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{scaffolder}
\alias{scaffolder}
\title{Scaffolding Interfaces to Packages in Other Programming Languages}
\description{
This package provides a comprehensive set of tools to scaffold
interfaces to modules, classes, functions, and
documentations written in other programming languages.
}
